# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Freedom.AI Operational Hub - API POC Viseu
  version: 0.1.0-poc
  description: >
    Contrato formal da API da POC. Escopo fechado para producao da POC:
    nao implementar alem do escopo contratado e nao deixar de implementar
    nenhuma funcionalidade do escopo.
servers:
  - url: http://localhost:3001
    description: Backend local
...
paths:
...
  /api/v1/audiencias:
    get:
      tags: [Audiencias]
      summary: Listar audiencias com filtros e paginacao
      parameters:
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/StatusAudiencia"
        - in: query
          name: trtId
          schema:
            type: string
            format: uuid
        - in: query
          name: prepostoId
          schema:
            type: string
            format: uuid
        - in: query
          name: parceiroId
          schema:
            type: string
            format: uuid
        - in: query
          name: modalidade
          schema:
            $ref: "#/components/schemas/Modalidade"
        - in: query
          name: dataInicio
          schema:
            type: string
            format: date
        - in: query
          name: dataFim
          schema:
            type: string
            format: date
        - in: query
          name: busca
          schema:
            type: string
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Lista paginada
          content:
            application/json:
              schema:
                type: object
                properties:
                  dados:
                    type: array
                    items:
                      $ref: "#/components/schemas/Audiencia"
                  paginacao:
                    $ref: "#/components/schemas/Paginacao"
                required: [dados, paginacao]
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
    post:
      tags: [Audiencias]
      summary: Criar audiencia manualmente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AudienciaCreateRequest"
      responses:
        "201":
          description: Audiencia criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audiencia"
        "400":
          $ref: "#/components/responses/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
  /api/v1/audiencias/kanban:
    get:
      tags: [Audiencias]
      summary: Kanban de audiencias por status
      responses:
        "200":
          description: Agrupamento por status
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: "#/components/schemas/Audiencia"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
  /api/v1/audiencias/dashboard:
    get:
      tags: [Audiencias]
      summary: KPIs operacionais
      responses:
        "200":
          description: Indicadores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
  /api/v1/audiencias/export:
    get:
      tags: [Audiencias]
      summary: Exportar audiencias em CSV/XLSX
      parameters:
        - in: query
          name: formato
          schema:
            type: string
            enum: [csv, xlsx]
            default: csv
      responses:
        "200":
          description: Arquivo exportado
          content:
            text/csv:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
  /api/v1/audiencias/{id}:
    get:
      tags: [Audiencias]
      summary: Detalhe da audiencia com timeline
      parameters:
        - $ref: "#/components/parameters/PathIdParam"
      responses:
        "200":
          description: Detalhe completo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AudienciaDetalhe"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
    patch:
      tags: [Audiencias]
      summary: Atualizar audiencia
      parameters:
        - $ref: "#/components/parameters/PathIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AudienciaUpdateRequest"
      responses:
        "200":
          description: Audiencia atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audiencia"
        "400":
          $ref: "#/components/responses/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
  /api/v1/audiencias/{id}/trocar-preposto:
    post:
      tags: [Audiencias]
      summary: Trocar preposto da audiencia
      parameters:
        - $ref: "#/components/parameters/PathIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prepostoNovoId:
                  type: string
                  format: uuid
                motivo:
                  type: string
              required: [prepostoNovoId, motivo]
      responses:
        "200":
          description: Preposto atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audiencia"
        "400":
          $ref: "#/components/responses/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
  /api/v1/audiencias/{id}/reenviar-confirmacao:
    post:
      tags: [Audiencias]
      summary: Reenviar confirmacao para preposto
      parameters:
        - $ref: "#/components/parameters/PathIdParam"
      responses:
        "200":
          description: Reenvio solicitado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  audienciaId:
                    type: string
                    format: uuid
                required: [message, audienciaId]
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
  /api/v1/audiencias/{id}/cancelar:
    post:
      tags: [Audiencias]
      summary: Cancelar audiencia manualmente
      parameters:
        - $ref: "#/components/parameters/PathIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                motivo:
                  type: string
              required: [motivo]
      responses:
        "200":
          description: Audiencia cancelada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audiencia"
        "400":
          $ref: "#/components/responses/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
  /api/v1/audiencias/{id}/confirmar-telefone:
    post:
      tags: [Audiencias]
      summary: Registrar confirmacao manual por telefone
      parameters:
        - $ref: "#/components/parameters/PathIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                observacao:
                  type: string
              required: [observacao]
      responses:
        "200":
          description: Confirmacao manual registrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audiencia"
        "400":
          $ref: "#/components/responses/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
  /api/v1/audiencias/{id}/check-in:
    post:
      tags: [Audiencias]
      summary: Registrar check-in no dia da audiencia
      parameters:
        - $ref: "#/components/parameters/PathIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                evento:
                  type: string
                  enum: [ESTOU_A_CAMINHO, JA_CHEGUEI, ESTOU_COM_PROBLEMA]
                observacao:
                  type: string
              required: [evento]
      responses:
        "200":
          description: Check-in registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audiencia"
        "400":
          $ref: "#/components/responses/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
  /api/v1/audiencias/{id}/relatorio:
    post:
      tags: [Audiencias]
      summary: Registrar relatorio pos-audiencia
      parameters:
        - $ref: "#/components/parameters/PathIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RelatorioCreateRequest"
      responses:
        "200":
          description: Relatorio salvo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelatorioAudiencia"
        "400":
          $ref: "#/components/responses/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathIdParam:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
  responses:
...
    ValidationErrorResponse:
      description: Erro de validacao
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
    UnauthorizedResponse:
      description: Nao autenticado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFoundResponse:
      description: Recurso nao encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    Modalidade:
      type: string
      enum: [PRESENCIAL, ONLINE]
    StatusAudiencia:
      type: string
      enum:
        - IMPORTADA
        - AGENDADA
        - A_CONFIRMAR
        - CONFIRMADA
        - NAO_POSSO
        - SEM_RESPOSTA
        - SUBSTITUICAO_NECESSARIA
        - EM_ANDAMENTO
        - CHECK_IN_PENDENTE
        - RELATORIO_PENDENTE
        - CONCLUIDA
        - CANCELADA
...
    OcorrenciaAudiencia:
      type: string
      enum: [SIM, NAO, REMARCADA]
    ResultadoAudiencia:
      type: string
      enum: [ACORDO, SEM_ACORDO, AUSENCIA, REDESIGNADA]
    DirecaoMensagem:
      type: string
      enum: [ENVIADA, RECEBIDA]
    TipoMensagem:
      type: string
      enum:
        - CONFIRMACAO_D1
        - REITERACAO_H1H30
        - CHECK_IN
        - RELATORIO_POS
        - SUBSTITUICAO_AVISO
        - ESCALONAMENTO
    StatusEnvioMensagem:
      type: string
      enum: [PENDENTE, ENVIADA, ENTREGUE, LIDA, FALHA]
    StatusSubstituicao:
      type: string
      enum: [ABERTA, RESOLVIDA, CANCELADA]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
        detalhes:
          type: array
          items:
            type: object
            additionalProperties: true
      required: [error]
    Paginacao:
      type: object
      properties:
        total:
          type: integer
        pagina:
          type: integer
        limite:
          type: integer
        totalPaginas:
          type: integer
      required: [total, pagina, limite, totalPaginas]
...
    Trt:
      type: object
      properties:
        id:
          type: string
          format: uuid
        numero:
          type: string
        nome:
          type: string
        ativo:
          type: boolean
      required: [id, numero, nome, ativo]
    Preposto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        telefoneWhatsapp:
          type: string
        email:
          type: string
          nullable: true
        cpf:
          type: string
          nullable: true
        ativo:
          type: boolean
      required: [id, nome, telefoneWhatsapp, ativo]
...
    Parceiro:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        ativo:
          type: boolean
      required: [id, nome, ativo]
...
    Audiencia:
      type: object
      properties:
        id:
          type: string
          format: uuid
        numeroProcesso:
          type: string
        reclamante:
          type: string
          nullable: true
        data:
          type: string
          format: date-time
        hora:
          type: string
          example: "14:30"
        modalidade:
          $ref: "#/components/schemas/Modalidade"
        local:
          type: string
          nullable: true
        link:
          type: string
          nullable: true
        trtId:
          type: string
          format: uuid
        vara:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/StatusAudiencia"
        prepostoId:
          type: string
          format: uuid
        parceiroId:
          type: string
          format: uuid
        importacaoId:
          type: string
          format: uuid
          nullable: true
        observacoes:
          type: string
          nullable: true
        trt:
          $ref: "#/components/schemas/Trt"
        preposto:
          $ref: "#/components/schemas/Preposto"
        parceiro:
          $ref: "#/components/schemas/Parceiro"
      required:
        - id
        - numeroProcesso
        - data
        - hora
        - modalidade
        - trtId
        - status
        - prepostoId
        - parceiroId
    AudienciaCreateRequest:
      type: object
      properties:
        numeroProcesso:
          type: string
        reclamante:
          type: string
        data:
          type: string
          format: date-time
        hora:
          type: string
        modalidade:
          $ref: "#/components/schemas/Modalidade"
        local:
          type: string
        link:
          type: string
        trtId:
          type: string
          format: uuid
        vara:
          type: string
        prepostoId:
          type: string
          format: uuid
        parceiroId:
          type: string
          format: uuid
        observacoes:
          type: string
      required:
        - numeroProcesso
        - data
        - hora
        - modalidade
        - trtId
        - prepostoId
        - parceiroId
    AudienciaUpdateRequest:
      type: object
      properties:
        numeroProcesso:
          type: string
        reclamante:
          type: string
        data:
          type: string
          format: date-time
        hora:
          type: string
        modalidade:
          $ref: "#/components/schemas/Modalidade"
        local:
          type: string
        link:
          type: string
        trtId:
          type: string
          format: uuid
        vara:
          type: string
        prepostoId:
          type: string
          format: uuid
        parceiroId:
          type: string
          format: uuid
        observacoes:
          type: string
        status:
          $ref: "#/components/schemas/StatusAudiencia"
      additionalProperties: false
    HistoricoStatus:
      type: object
      properties:
        id:
          type: string
          format: uuid
        audienciaId:
          type: string
          format: uuid
        statusAnterior:
          $ref: "#/components/schemas/StatusAudiencia"
        statusNovo:
          $ref: "#/components/schemas/StatusAudiencia"
        motivo:
          type: string
          nullable: true
        atualizadoPor:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, audienciaId, statusAnterior, statusNovo, atualizadoPor, createdAt]
    Mensagem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        audienciaId:
          type: string
          format: uuid
        tipo:
          $ref: "#/components/schemas/TipoMensagem"
        direcao:
          $ref: "#/components/schemas/DirecaoMensagem"
        conteudo:
          type: string
        respostaBotao:
          type: string
          nullable: true
        observacao:
          type: string
          nullable: true
        statusEnvio:
          $ref: "#/components/schemas/StatusEnvioMensagem"
        createdAt:
          type: string
          format: date-time
      required: [id, audienciaId, tipo, direcao, conteudo, statusEnvio, createdAt]
    Substituicao:
      type: object
      properties:
        id:
          type: string
          format: uuid
        audienciaId:
          type: string
          format: uuid
        prepostoAnteriorId:
          type: string
          format: uuid
        prepostoNovoId:
          type: string
          format: uuid
          nullable: true
        motivo:
          type: string
        status:
          $ref: "#/components/schemas/StatusSubstituicao"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - audienciaId
        - prepostoAnteriorId
        - motivo
        - status
        - createdAt
    RelatorioAudiencia:
      type: object
      properties:
        id:
          type: string
          format: uuid
        audienciaId:
          type: string
          format: uuid
        audienciaOcorreu:
          $ref: "#/components/schemas/OcorrenciaAudiencia"
        resultado:
          $ref: "#/components/schemas/ResultadoAudiencia"
        advogadoPresente:
          type: boolean
        advogadoDominioCaso:
          type: boolean
        problemaRelevante:
          type: boolean
        relato:
          type: string
          nullable: true
      required: [id, audienciaId]
    RelatorioCreateRequest:
      type: object
      properties:
        audienciaOcorreu:
          $ref: "#/components/schemas/OcorrenciaAudiencia"
        resultado:
          $ref: "#/components/schemas/ResultadoAudiencia"
        advogadoPresente:
          type: boolean
        advogadoDominioCaso:
          type: boolean
        problemaRelevante:
          type: boolean
        relato:
          type: string
      required:
        - audienciaOcorreu
        - resultado
        - advogadoPresente
        - advogadoDominioCaso
        - problemaRelevante
    AudienciaDetalhe:
      allOf:
        - $ref: "#/components/schemas/Audiencia"
        - type: object
          properties:
            historicoStatus:
              type: array
              items:
                $ref: "#/components/schemas/HistoricoStatus"
            mensagens:
              type: array
              items:
                $ref: "#/components/schemas/Mensagem"
            substituicoes:
              type: array
              items:
                $ref: "#/components/schemas/Substituicao"
            relatorio:
              $ref: "#/components/schemas/RelatorioAudiencia"
    DashboardResponse:
      type: object
      properties:
        totalAtivas:
          type: integer
        audienciasHoje:
          type: integer
        audienciasSemana:
          type: integer
        aguardandoConfirmacao:
          type: integer
        semResposta:
          type: integer
        porStatus:
          type: array
          items:
            type: object
            properties:
              status:
                $ref: "#/components/schemas/StatusAudiencia"
              total:
                type: integer
            required: [status, total]
      required:
        - totalAtivas
        - audienciasHoje
        - audienciasSemana
        - aguardandoConfirmacao
        - semResposta
        - porStatus
...
security:
  - bearerAuth: []